requirements.txt
 python-ffmpeg
audiosegment